requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.10.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
huggingface-hub==0.13.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Levenshtein==0.20.9
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.4.1
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn-pandas==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.26.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
